+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 
-- added `F mode` prompt option to change the safe search filter (e.g. `F moderate`)
-- added `e engine` prompt option to add or remove engines from the interactive console
+- added `-q/--query` option as explicit alternative to the default positional query
+- added `F mode` prompt option to change the safe search filter in the interactive console (e.g. `F moderate`)
+- added `e engine` prompt option to add or remove engines for the current search in the interactive console
 
 ## v0.6.0